Added JWT auth and role model

+++ src/project/api/client_routes.py
@@ -1,6 +1,8 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
-from project.api.depends import database, client_repo
+from project.schemas.user import UserSchema
+
+from project.api.depends import database, client_repo, get_current_user, check_for_admin_access
 from project.schemas.client import *
 from project.core.exceptions import *
 
@@ -33,7 +35,9 @@ async def get_client_by_id(client_id: int) -> ClientSchema:
 
 
 @client_router.post("/add_client", response_model=ClientSchema, status_code=status.HTTP_201_CREATED)
-async def add_client(client_dto: ClientCreateUpdateSchema) -> ClientSchema:
+async def add_client(client_dto: ClientCreateUpdateSchema,
+                     current_user: UserSchema = Depends(get_current_user), ) -> ClientSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             new_client = await client_repo.create_client(session=session, client=client_dto)
@@ -47,7 +51,9 @@ async def add_client(client_dto: ClientCreateUpdateSchema) -> ClientSchema:
     response_model=ClientSchema,
     status_code=status.HTTP_200_OK,
 )
-async def update_client(client_id: int, client_dto: ClientCreateUpdateSchema) -> ClientSchema:
+async def update_client(client_id: int, client_dto: ClientCreateUpdateSchema,
+                        current_user: UserSchema = Depends(get_current_user), ) -> ClientSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             updated_client = await client_repo.update_client(
@@ -61,7 +67,8 @@ async def update_client(client_id: int, client_dto: ClientCreateUpdateSchema) ->
 
 
 @client_router.delete("/delete_client/{client_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_client(client_id: int) -> None:
+async def delete_client(client_id: int, current_user: UserSchema = Depends(get_current_user), ) -> None:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             client = await client_repo.delete_client(session=session, client_id=client_id)
