+++ src/project/api/expense_routes.py
@@ -1,6 +1,8 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
-from project.api.depends import database, expense_repo
+from project.schemas.user import UserSchema
+
+from project.api.depends import database, expense_repo, get_current_user, check_for_admin_access
 from project.schemas.expense import *
 from project.core.exceptions import *
 
@@ -33,7 +35,9 @@ async def get_expense_by_id(expense_id: int) -> ExpenseSchema:
 
 
 @expense_router.post("/add_expense", response_model=ExpenseSchema, status_code=status.HTTP_201_CREATED)
-async def add_expense(expense_dto: ExpenseCreateUpdateSchema) -> ExpenseSchema:
+async def add_expense(expense_dto: ExpenseCreateUpdateSchema,
+                      current_user: UserSchema = Depends(get_current_user), ) -> ExpenseSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             new_expense = await expense_repo.create_expense(session=session, expense=expense_dto)
@@ -47,7 +51,9 @@ async def add_expense(expense_dto: ExpenseCreateUpdateSchema) -> ExpenseSchema:
     response_model=ExpenseSchema,
     status_code=status.HTTP_200_OK,
 )
-async def update_expense(expense_id: int, expense_dto: ExpenseCreateUpdateSchema) -> ExpenseSchema:
+async def update_expense(expense_id: int, expense_dto: ExpenseCreateUpdateSchema,
+                         current_user: UserSchema = Depends(get_current_user), ) -> ExpenseSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             updated_expense = await expense_repo.update_expense(
@@ -61,7 +67,8 @@ async def update_expense(expense_id: int, expense_dto: ExpenseCreateUpdateSchema
 
 
 @expense_router.delete("/delete_expense/{expense_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_expense(expense_id: int) -> None:
+async def delete_expense(expense_id: int, current_user: UserSchema = Depends(get_current_user),) -> None:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             expense = await expense_repo.delete_expense(session=session, expense_id=expense_id)
