+++ src/project/api/good_routes.py
@@ -1,6 +1,8 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
-from project.api.depends import database, good_repo
+from project.schemas.user import UserSchema
+
+from project.api.depends import database, good_repo, get_current_user, check_for_admin_access
 from project.schemas.good import *
 from project.core.exceptions import *
 
@@ -33,7 +35,8 @@ async def get_good_by_id(good_id: int) -> GoodSchema:
 
 
 @good_router.post("/add_good", response_model=GoodSchema, status_code=status.HTTP_201_CREATED)
-async def add_good(good_dto: GoodCreateUpdateSchema) -> GoodSchema:
+async def add_good(good_dto: GoodCreateUpdateSchema, current_user: UserSchema = Depends(get_current_user),) -> GoodSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             new_good = await good_repo.create_good(session=session, good=good_dto)
@@ -47,7 +50,8 @@ async def add_good(good_dto: GoodCreateUpdateSchema) -> GoodSchema:
     response_model=GoodSchema,
     status_code=status.HTTP_200_OK,
 )
-async def update_good(good_id: int, good_dto: GoodCreateUpdateSchema) -> GoodSchema:
+async def update_good(good_id: int, good_dto: GoodCreateUpdateSchema, current_user: UserSchema = Depends(get_current_user),) -> GoodSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             updated_good = await good_repo.update_good(
@@ -61,7 +65,8 @@ async def update_good(good_id: int, good_dto: GoodCreateUpdateSchema) -> GoodSch
 
 
 @good_router.delete("/delete_good/{good_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_good(good_id: int) -> None:
+async def delete_good(good_id: int, current_user: UserSchema = Depends(get_current_user),) -> None:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             good = await good_repo.delete_good(session=session, good_id=good_id)
