+++ src/project/api/storagemethod_routes.py
@@ -1,6 +1,8 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
-from project.api.depends import database, storage_method_repo
+from project.schemas.user import UserSchema
+
+from project.api.depends import database, storage_method_repo, get_current_user, check_for_admin_access
 from project.schemas.storagemethod import *
 from project.core.exceptions import *
 
@@ -35,7 +37,9 @@ async def get_storage_method_by_id(storage_method_id: int) -> StorageMethodSchem
 
 @storage_method_router.post("/add_storage_method", response_model=StorageMethodSchema,
                             status_code=status.HTTP_201_CREATED)
-async def add_storage_method(storage_method_dto: StorageMethodCreateUpdateSchema) -> StorageMethodSchema:
+async def add_storage_method(storage_method_dto: StorageMethodCreateUpdateSchema,
+                             current_user: UserSchema = Depends(get_current_user), ) -> StorageMethodSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             new_storage_method = await storage_method_repo.create_storage_method(session=session,
@@ -51,7 +55,9 @@ async def add_storage_method(storage_method_dto: StorageMethodCreateUpdateSchema
     status_code=status.HTTP_200_OK,
 )
 async def update_storage_method(storage_method_id: int,
-                                storage_method_dto: StorageMethodCreateUpdateSchema) -> StorageMethodSchema:
+                                storage_method_dto: StorageMethodCreateUpdateSchema,
+                                current_user: UserSchema = Depends(get_current_user), ) -> StorageMethodSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             updated_storage_method = await storage_method_repo.update_storage_method(
@@ -65,7 +71,8 @@ async def update_storage_method(storage_method_id: int,
 
 
 @storage_method_router.delete("/delete_storage_method/{storage_method_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_storage_method(storage_method_id: int) -> None:
+async def delete_storage_method(storage_method_id: int, current_user: UserSchema = Depends(get_current_user), ) -> None:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             storagemethod = await storage_method_repo.delete_storage_method(session=session,
