+++ src/project/api/supplier_routes.py
@@ -1,6 +1,8 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
-from project.api.depends import database, supplier_repo
+from project.schemas.user import UserSchema
+
+from project.api.depends import database, supplier_repo, get_current_user, check_for_admin_access
 from project.schemas.supplier import *
 from project.core.exceptions import *
 
@@ -33,7 +35,9 @@ async def get_supplier_by_id(supplier_id: int) -> SupplierSchema:
 
 
 @supplier_router.post("/add_supplier", response_model=SupplierSchema, status_code=status.HTTP_201_CREATED)
-async def add_supplier(supplier_dto: SupplierCreateUpdateSchema) -> SupplierSchema:
+async def add_supplier(supplier_dto: SupplierCreateUpdateSchema,
+                       current_user: UserSchema = Depends(get_current_user), ) -> SupplierSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             new_supplier = await supplier_repo.create_supplier(session=session, supplier=supplier_dto)
@@ -47,7 +51,9 @@ async def add_supplier(supplier_dto: SupplierCreateUpdateSchema) -> SupplierSche
     response_model=SupplierSchema,
     status_code=status.HTTP_200_OK,
 )
-async def update_supplier(supplier_id: int, supplier_dto: SupplierCreateUpdateSchema) -> SupplierSchema:
+async def update_supplier(supplier_id: int, supplier_dto: SupplierCreateUpdateSchema,
+                          current_user: UserSchema = Depends(get_current_user), ) -> SupplierSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             updated_supplier = await supplier_repo.update_supplier(
@@ -61,7 +67,8 @@ async def update_supplier(supplier_id: int, supplier_dto: SupplierCreateUpdateSc
 
 
 @supplier_router.delete("/delete_supplier/{supplier_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_supplier(supplier_id: int) -> None:
+async def delete_supplier(supplier_id: int, current_user: UserSchema = Depends(get_current_user), ) -> None:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             supplier = await supplier_repo.delete_supplier(session=session, supplier_id=supplier_id)
